+++ src/main/java/ru/itis/grant/dao/interfaces/EventDao.java
@@ -31,7 +31,7 @@ public interface EventDao {
 
     List<Event> getUserEvents(long userId);
 
-    List<Event> getExpertEvents(String token, long from, long count);
+    List<Event> getExpertEvents(String token, int from, int count);
 
     boolean eventExistenceById(long id);
 
@@ -41,7 +41,7 @@ public interface EventDao {
 
     boolean organizerEventExistence(long eventId, String token);
 
-    List<Event> getOrganizerEvents(String token, long from, long count);
+    List<Event> getOrganizerEvents(String token, int from, int count);
 
     void addExpertToEvent(long eventId, long expertId);
